requirements.txt
torch>=2.0.0
torchvision>=0.15.0
albumentations>=1.3.0
pytorch-grad-cam>=1.4.0
thop>=0.1.1
onnx>=1.14.0
onnxruntime>=1.15.0
opencv-python>=4.7.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
Pillow>=9.5.0
PyYAML>=6.0
tqdm>=4.65.0
jupyter>=1.0.0
